# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Constructatech Ventures API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /portal/dashboard:
    get:
      operationId: getPortalDashboard
      tags: [portal]
      summary: Get portal dashboard summary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalDashboard"
        "401":
          description: Unauthorized
  /portal/projects:
    get:
      operationId: listPortalProjects
      tags: [portal]
      summary: List my projects
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
  /portal/projects/{id}:
    get:
      operationId: getPortalProject
      tags: [portal]
      summary: Get project detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /portal/tickets:
    get:
      operationId: listPortalTickets
      tags: [portal]
      summary: List my tickets
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
        "401":
          description: Unauthorized
    post:
      operationId: createPortalTicket
      tags: [portal]
      summary: Create a new support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "401":
          description: Unauthorized
  /portal/tickets/{id}:
    get:
      operationId: getPortalTicket
      tags: [portal]
      summary: Get ticket with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /portal/tickets/{id}/messages:
    post:
      operationId: addTicketMessage
      tags: [portal]
      summary: Add message to ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketMessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketMessage"
        "401":
          description: Unauthorized
  /portal/invoices:
    get:
      operationId: listPortalInvoices
      tags: [portal]
      summary: List my invoices
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
        "401":
          description: Unauthorized
  /portal/account:
    get:
      operationId: getPortalAccount
      tags: [portal]
      summary: Get my account details
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerAccount"
        "401":
          description: Unauthorized
    patch:
      operationId: updatePortalAccount
      tags: [portal]
      summary: Update my account details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerAccountUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerAccount"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    PortalDashboard:
      type: object
      required: [activeProjects, openTickets, unpaidInvoices, recentActivity]
      properties:
        activeProjects:
          type: number
        openTickets:
          type: number
        unpaidInvoices:
          type: number
        totalInvoiceAmount:
          type: number
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
    ActivityItem:
      type: object
      required: [id, type, description, createdAt]
      properties:
        id:
          type: number
        type:
          type: string
        description:
          type: string
        createdAt:
          type: string
    Milestone:
      type: object
      required: [label, done]
      properties:
        label:
          type: string
        dueDate:
          type: ["string", "null"]
        done:
          type: boolean
    Project:
      type: object
      required: [id, title, description, status, customerId, milestones, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [scoping, in-progress, on-hold, completed]
        customerId:
          type: number
        customerName:
          type: ["string", "null"]
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/Milestone"
        startDate:
          type: ["string", "null"]
        targetDate:
          type: ["string", "null"]
        createdAt:
          type: string
...
    TicketMessage:
      type: object
      required: [id, author, body, isStaff, createdAt]
      properties:
        id:
          type: number
        author:
          type: string
        body:
          type: string
        isStaff:
          type: boolean
        createdAt:
          type: string
    Ticket:
      type: object
      required: [id, subject, type, priority, status, customerId, messages, createdAt]
      properties:
        id:
          type: number
        subject:
          type: string
        type:
          type: string
          enum: [support, change-request]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [open, in-progress, resolved, closed]
        customerId:
          type: number
        customerName:
          type: ["string", "null"]
        projectId:
          type: ["number", "null"]
        messages:
          type: array
          items:
            $ref: "#/components/schemas/TicketMessage"
        createdAt:
          type: string
    TicketInput:
      type: object
      required: [subject, type, priority, body]
      properties:
        subject:
          type: string
          minLength: 1
        type:
          type: string
          enum: [support, change-request]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        projectId:
          type: ["number", "null"]
        body:
          type: string
          minLength: 1
    TicketMessageInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
    Invoice:
      type: object
      required: [id, number, customerId, amount, currency, status, issueDate, dueDate, createdAt]
      properties:
        id:
          type: number
        number:
          type: string
        customerId:
          type: number
        customerName:
          type: ["string", "null"]
        projectId:
          type: ["number", "null"]
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [draft, sent, paid, overdue]
        issueDate:
          type: string
        dueDate:
          type: string
        createdAt:
          type: string
...
    CustomerAccount:
      type: object
      required: [id, companyName, contactName, email, phone]
      properties:
        id:
          type: number
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
    CustomerAccountUpdate:
      type: object
      properties:
        contactName:
          type: string
        phone:
          type: string
        companyName:
          type: string