# lib/api-zod/src/generated/types/customerAccount.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Constructatech Ventures API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerAccount {
  id: number;
  companyName: string;
  contactName: string;
  email: string;
  phone: string;
}