# lib/api-zod/src/generated/types/portalDashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Constructatech Ventures API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from './activityItem';

export interface PortalDashboard {
  activeProjects: number;
  openTickets: number;
  unpaidInvoices: number;
  totalInvoiceAmount?: number;
  recentActivity: ActivityItem[];
}