# lib/api-zod/src/generated/types/ticketMessageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Constructatech Ventures API
 * OpenAPI spec version: 0.1.0
 */

export interface TicketMessageInput {
  /** @minLength 1 */
  body: string;
}